Deep Lake upgrade to include attribute search, distance metrics, returning scores and MMR (#2455)
### Features include

- Metadata based embedding search
- Choice of distance metric function (`L2` for Euclidean, `L1` for
Nuclear, `max` L-infinity distance, `cos` for cosine similarity, 'dot'
for dot product. Defaults to `L2`
- Returning scores
- Max Marginal Relevance Search
- Deleting samples from the dataset

### Notes
- Added numerous tests, let me know if you would like to shorten them or
make smarter

---------

+++ langchain/vectorstores/deeplake.py
@@ -3,40 +3,76 @@ from __future__ import annotations
 
 import logging
 import uuid
-from typing import Any, Iterable, List, Optional, Sequence
+from functools import partial
+from typing import Any, Dict, Iterable, List, Optional, Sequence, Tuple
 
 import numpy as np
 
 from langchain.docstore.document import Document
 from langchain.embeddings.base import Embeddings
 from langchain.vectorstores.base import VectorStore
+from langchain.vectorstores.utils import maximal_marginal_relevance
 
 logger = logging.getLogger()
 
+distance_metric_map = {
+    "l2": lambda a, b: np.linalg.norm(a - b, axis=1, ord=2),
+    "l1": lambda a, b: np.linalg.norm(a - b, axis=1, ord=1),
+    "max": lambda a, b: np.linalg.norm(a - b, axis=1, ord=np.inf),
+    "cos": lambda a, b: np.dot(a, b.T)
+    / (np.linalg.norm(a) * np.linalg.norm(b, axis=1)),
+    "dot": lambda a, b: np.dot(a, b.T),
+}
 
-def L2_search(
-    query_embedding: np.ndarray, data_vectors: np.ndarray, k: int = 4
-) -> list:
-    """naive L2 search for nearest neighbors"""
-    # Calculate the L2 distance between the query_vector and all data_vectors
-    distances = np.linalg.norm(data_vectors - query_embedding, axis=1)
 
-    # Sort the distances and return the indices of the k nearest vectors
-    nearest_indices = np.argsort(distances)[:k]
-    return nearest_indices.tolist()
+def vector_search(
+    query_embedding: np.ndarray,
+    data_vectors: np.ndarray,
+    distance_metric: str = "L2",
+    k: Optional[int] = 4,
+) -> Tuple[List, List]:
+    """Naive search for nearest neighbors
+
+    args:
+        query_embedding: np.ndarray
+        data_vectors: np.ndarray
+        k (int): number of nearest neighbors
+        distance_metric: distance function 'L2' for Euclidean, 'L1' for Nuclear, 'Max'
+            l-infinity distnace, 'cos' for cosine similarity, 'dot' for dot product
+
+    returns:
+        nearest_indices: List, indices of nearest neighbors
+    """
+    # Calculate the distance between the query_vector and all data_vectors
+    distances = distance_metric_map[distance_metric](query_embedding, data_vectors)
+    nearest_indices = np.argsort(distances)
+    nearest_indices = (
+        nearest_indices[::-1][:k] if distance_metric in ["cos"] else nearest_indices[:k]
+    )
+
+    return nearest_indices.tolist(), distances[nearest_indices].tolist()
+
+
+def dp_filter(x: dict, filter: Dict[str, str]) -> bool:
+    """Filter helper function for Deep Lake"""
+    metadata = x["metadata"].data()["value"]
+    return all(k in metadata and v == metadata[k] for k, v in filter.items())
 
 
 class DeepLake(VectorStore):
     """Wrapper around Deep Lake, a data lake for deep learning applications.
 
-    It not only stores embeddings, but also the original data and queries with
-    version control automatically enabled.
+    We implement naive similarity search and filtering for fast prototyping,
+    but it can be extended with Tensor Query Language (TQL) for production use cases
+    over billion rows.
 
-    It is more than just a vector store. You can use the dataset to fine-tune
-    your own LLM models or use it for other downstream tasks.
+    Why Deep Lake?
 
-    We implement naive similiarity search, but it can be extended with Tensor
-    Query Language (TQL for production use cases) over billion rows.
+    - Not only stores embeddings, but also the original data with version control.
+    - Serverless, doesn't require another service and can be used with major
+        cloud providers (S3, GCS, etc.)
+    - More than just a multi-modal vector store. You can use the dataset
+        to fine-tune your own LLM models.
 
     To use, you should have the ``deeplake`` python package installed.
 
@@ -80,10 +116,34 @@ class DeepLake(VectorStore):
         else:
             self.ds = deeplake.empty(dataset_path, token=token, overwrite=True)
             with self.ds:
-                self.ds.create_tensor("text", htype="text")
-                self.ds.create_tensor("metadata", htype="json")
-                self.ds.create_tensor("embedding", htype="generic")
-                self.ds.create_tensor("ids", htype="text")
+                self.ds.create_tensor(
+                    "text",
+                    htype="text",
+                    create_id_tensor=False,
+                    create_sample_info_tensor=False,
+                    create_shape_tensor=False,
+                )
+                self.ds.create_tensor(
+                    "metadata",
+                    htype="json",
+                    create_id_tensor=False,
+                    create_sample_info_tensor=False,
+                    create_shape_tensor=False,
+                )
+                self.ds.create_tensor(
+                    "embedding",
+                    htype="generic",
+                    create_id_tensor=False,
+                    create_sample_info_tensor=False,
+                    create_shape_tensor=False,
+                )
+                self.ds.create_tensor(
+                    "ids",
+                    htype="text",
+                    create_id_tensor=False,
+                    create_sample_info_tensor=False,
+                    create_shape_tensor=False,
+                )
 
         self._embedding_function = embedding_function
 
@@ -116,11 +176,9 @@ class DeepLake(VectorStore):
             embeddings = self._embedding_function.embed_documents(text_list)
 
         if metadatas is None:
-            metadatas_to_use: Sequence[Optional[dict]] = [None] * len(text_list)
-        else:
-            metadatas_to_use = metadatas
+            metadatas = [{}] * len(text_list)
 
-        elements = zip(text_list, embeddings, metadatas_to_use, ids)
+        elements = zip(text_list, embeddings, metadatas, ids)
 
         @self._deeplake.compute
         def ingest(sample_in: list, sample_out: list) -> None:
@@ -133,31 +191,209 @@ class DeepLake(VectorStore):
             sample_out.append(s)
 
         ingest().eval(list(elements), self.ds)
-        self.ds.commit()
+        self.ds.commit(allow_empty=True)
 
         return ids
 
-    def similarity_search(
-        self, query: str, k: int = 4, **kwargs: Any
-    ) -> List[Document]:
-        """Return docs most similar to query."""
+    def search(
+        self,
+        query: Any[str, None] = None,
+        embedding: Any[float, None] = None,
+        k: int = 4,
+        distance_metric: str = "L2",
+        use_maximal_marginal_relevance: Optional[bool] = False,
+        fetch_k: Optional[int] = 20,
+        filter: Optional[Dict[str, str]] = None,
+        return_score: Optional[bool] = False,
+        **kwargs: Any,
+    ) -> Any[List[Document], List[Tuple[Document, float]]]:
+        """Return docs most similar to query.
+
+        Args:
+            query: Text to look up documents similar to.
+            embedding: Embedding function to use. Defaults to None.
+            k: Number of Documents to return. Defaults to 4.
+            distance_metric: `L2` for Euclidean, `L1` for Nuclear,
+                `max` L-infinity distance, `cos` for cosine similarity,
+                'dot' for dot product. Defaults to `L2`.
+            filter: Attribute filter by metadata example {'key': 'value'}.
+                Defaults to None.
+            maximal_marginal_relevance: Whether to use maximal marginal relevance.
+                Defaults to False.
+            fetch_k: Number of Documents to fetch to pass to MMR algorithm.
+                Defaults to 20.
+            return_score: Whether to return the score. Defaults to False.
+
+        Returns:
+            List of Documents selected by the specified distance metric,
+            if return_score True, return a tuple of (Document, score)
+        """
+        view = self.ds
+
+        # attribute based filtering
+        if filter is not None:
+            view = view.filter(partial(dp_filter, filter=filter))
+
+            if len(view) == 0:
+                return []
+
         if self._embedding_function is None:
-            self.ds.summary()
-            ds_view = self.ds.filter(lambda x: query in x["text"].data()["value"])
+            view = view.filter(lambda x: query in x["text"].data()["value"])
+            scores = [1.0] * len(view)
+
+            if use_maximal_marginal_relevance:
+                raise ValueError(
+                    "For MMR search, you must specify an embedding function on"
+                    "creation."
+                )
+
         else:
-            query_emb = np.array(self._embedding_function.embed_query(query))
-            embeddings = self.ds.embedding.numpy()
-            indices = L2_search(query_emb, embeddings, k=k)
-            ds_view = self.ds[indices]
+            emb = embedding or self._embedding_function.embed_query(
+                query
+            )  # type: ignore
+            query_emb = np.array(emb, dtype=np.float32)
+            embeddings = view.embedding.numpy()
+
+            k_search = fetch_k if use_maximal_marginal_relevance else k
+            indices, scores = vector_search(
+                query_emb,
+                embeddings,
+                k=k_search,
+                distance_metric=distance_metric.lower(),
+            )
+            view = view[indices]
+
+            if use_maximal_marginal_relevance:
+                indices = maximal_marginal_relevance(
+                    query_emb, embeddings[indices], k=min(k, len(indices))
+                )
+                view = view[indices]
+                scores = [scores[i] for i in indices]
 
         docs = [
             Document(
                 page_content=el["text"].data()["value"],
                 metadata=el["metadata"].data()["value"],
             )
-            for el in ds_view
+            for el in view
         ]
+
+        if return_score:
+            return [(doc, score) for doc, score in zip(docs, scores)]
+
         return docs
+
+    def similarity_search(
+        self, query: str, k: int = 4, **kwargs: Any
+    ) -> List[Document]:
+        """Return docs most similar to query.
+
+        Args:
+            query: text to embed and run the query on.
+            k: Number of Documents to return.
+                Defaults to 4.
+            query: Text to look up documents similar to.
+            embedding: Embedding function to use.
+                Defaults to None.
+            k: Number of Documents to return.
+                Defaults to 4.
+            distance_metric: `L2` for Euclidean, `L1` for Nuclear, `max`
+                L-infinity distance, `cos` for cosine similarity, 'dot' for dot product
+                Defaults to `L2`.
+            filter: Attribute filter by metadata example {'key': 'value'}.
+                Defaults to None.
+            maximal_marginal_relevance: Whether to use maximal marginal relevance.
+                Defaults to False.
+            fetch_k: Number of Documents to fetch to pass to MMR algorithm.
+                Defaults to 20.
+            return_score: Whether to return the score. Defaults to False.
+
+        Returns:
+            List of Documents most similar to the query vector.
+        """
+        return self.search(query=query, k=k, **kwargs)
+
+    def similarity_search_by_vector(
+        self, embedding: List[float], k: int = 4, **kwargs: Any
+    ) -> List[Document]:
+        """Return docs most similar to embedding vector.
+
+        Args:
+            embedding: Embedding to look up documents similar to.
+            k: Number of Documents to return. Defaults to 4.
+        Returns:
+            List of Documents most similar to the query vector.
+        """
+        return self.search(embedding=embedding, k=k, **kwargs)
+
+    def similarity_search_with_score(
+        self,
+        query: str,
+        distance_metric: str = "L2",
+        k: int = 4,
+        filter: Optional[Dict[str, str]] = None,
+    ) -> List[Tuple[Document, float]]:
+        """Run similarity search with Deep Lake with distance returned.
+
+        Args:
+            query (str): Query text to search for.
+            distance_metric: `L2` for Euclidean, `L1` for Nuclear, `max` L-infinity
+                distance, `cos` for cosine similarity, 'dot' for dot product.
+                Defaults to `L2`.
+            k (int): Number of results to return. Defaults to 4.
+            filter (Optional[Dict[str, str]]): Filter by metadata. Defaults to None.
+        Returns:
+            List[Tuple[Document, float]]: List of documents most similar to the query
+                text with distance in float.
+        """
+        return self.search(
+            query=query,
+            k=k,
+            filter=filter,
+            return_score=True,
+            distance_metric=distance_metric,
+        )
+
+    def max_marginal_relevance_search_by_vector(
+        self, embedding: List[float], k: int = 4, fetch_k: int = 20
+    ) -> List[Document]:
+        """Return docs selected using the maximal marginal relevance.
+        Maximal marginal relevance optimizes for similarity to query AND diversity
+        among selected documents.
+        Args:
+            embedding: Embedding to look up documents similar to.
+            k: Number of Documents to return. Defaults to 4.
+            fetch_k: Number of Documents to fetch to pass to MMR algorithm.
+        Returns:
+            List of Documents selected by maximal marginal relevance.
+        """
+        return self.search(
+            embedding=embedding,
+            k=k,
+            fetch_k=fetch_k,
+            use_maximal_marginal_relevance=True,
+        )
+
+    def max_marginal_relevance_search(
+        self, query: str, k: int = 4, fetch_k: int = 20
+    ) -> List[Document]:
+        """Return docs selected using the maximal marginal relevance.
+        Maximal marginal relevance optimizes for similarity to query AND diversity
+        among selected documents.
+        Args:
+            query: Text to look up documents similar to.
+            k: Number of Documents to return. Defaults to 4.
+            fetch_k: Number of Documents to fetch to pass to MMR algorithm.
+        Returns:
+            List of Documents selected by maximal marginal relevance.
+        """
+        if self._embedding_function is None:
+            raise ValueError(
+                "For MMR search, you must specify an embedding function on" "creation."
+            )
+        return self.search(
+            query=query, k=k, fetch_k=fetch_k, use_maximal_marginal_relevance=True
+        )
 
     @classmethod
     def from_texts(
@@ -171,22 +407,24 @@ class DeepLake(VectorStore):
     ) -> DeepLake:
         """Create a Deep Lake dataset from a raw documents.
 
-        If a persist_directory is specified, the collection will be persisted there.
+        If a dataset_path is specified, the dataset will be persisted there.
         Otherwise, the data will be ephemeral in-memory.
 
         Args:
             path (str, pathlib.Path): - The full path to the dataset. Can be:
-                - a Deep Lake cloud path of the form ``hub://username/datasetname``.
+                - Deep Lake cloud path of the form ``hub://username/dataset_name``.
                     To write to Deep Lake cloud datasets,
                     ensure that you are logged in to Deep Lake
                     (use 'activeloop login' from command line)
-                - an s3 path of the form ``s3://bucketname/path/to/dataset``.
-                    Credentials are required in either the environment or
-                    passed to the creds argument.
-                - a local file system path of the form ``./path/to/dataset`` or
+                - AWS S3 path of the form ``s3://bucketname/path/to/dataset``.
+                    Credentials are required in either the environment
+                - Google Cloud Storage path of the form
+                    ``gcs://bucketname/path/to/dataset``Credentials are required
+                    in either the environment
+                - Local file system path of the form ``./path/to/dataset`` or
                     ``~/path/to/dataset`` or ``path/to/dataset``.
-                - a memory path of the form ``mem://path/to/dataset`` which doesn't
-                    save the dataset but keeps it in memory instead.
+                - In-memory path of the form ``mem://path/to/dataset`` which doesn't
+                    save the dataset, but keeps it in memory instead.
                     Should be used only for testing as it does not persist.
             documents (List[Document]): List of documents to add.
             embedding (Optional[Embeddings]): Embedding function. Defaults to None.
@@ -203,9 +441,48 @@ class DeepLake(VectorStore):
         deeplake_dataset.add_texts(texts=texts, metadatas=metadatas, ids=ids)
         return deeplake_dataset
 
+    def delete(
+        self,
+        ids: Any[List[str], None] = None,
+        filter: Any[Dict[str, str], None] = None,
+        delete_all: Any[bool, None] = None,
+    ) -> bool:
+        """Delete the entities in the dataset
+
+        Args:
+            ids (Optional[List[str]], optional): The document_ids to delete.
+                Defaults to None.
+            filter (Optional[Dict[str, str]], optional): The filter to delete by.
+                Defaults to None.
+            delete_all (Optional[bool], optional): Whether to drop the dataset.
+                Defaults to None.
+        """
+        if delete_all:
+            self.ds.delete()
+            return True
+
+        view = None
+        if ids:
+            view = self.ds.filter(lambda x: x["ids"].data()["value"] in ids)
+            ids = list(view.sample_indices)
+
+        if filter:
+            if view is None:
+                view = self.ds
+            view = view.filter(partial(dp_filter, filter=filter))
+            ids = list(view.sample_indices)
+
+        with self.ds:
+            for id in sorted(ids)[::-1]:
+                self.ds.pop(id)
+
+            self.ds.commit(f"deleted {len(ids)} samples", allow_empty=True)
+
+        return True
+
     def delete_dataset(self) -> None:
         """Delete the collection."""
-        self.ds.delete()
+        self.delete(delete_all=True)
 
     def persist(self) -> None:
         """Persist the collection."""
